moved app to its own file
